Refactor: Service peek -> map 변경, test code directory 변경
